Use BS ISO 3166-2:2013 GB codes for la-nir
The old ones seem to have been made-up.  These are for the old 26 local
authorities that were reorganised in 2015.

Also removed the start-date of Belfast, since the code didn't change with the
reorganisation.  I searched for a start-date in legislation, but didn't find
one.  Up to the custodian to supply if available.